+++ Tests/LegacyUnit/SchedulerTasks/MailNotifierTest.php
@@ -106,7 +106,6 @@ final class MailNotifierTest extends TestCase
         Bootstrap::initializeBackendAuthentication();
 
         if (Typo3Version::isAtLeast(10)) {
-            // @phpstan-ignore-next-line This line is for TYPO3 10LTS only, and we currently are on 9LTS.
             $this->languageService = LanguageService::create('default');
         } else {
             $this->languageService = new LanguageService();